update SDK v3 move configs
