+++ chat_sample/lib/main.dart
@@ -14,12 +14,12 @@ import 'firebase_options.dart';
 import 'src/chat_details_screen.dart';
 import 'src/chat_dialog_screen.dart';
 import 'src/chat_dialog_resizable_screen.dart';
+import 'src/config.dart' as config;
 import 'src/login_screen.dart';
 import 'src/managers/push_notifications_manager.dart';
 import 'src/select_dialog_screen.dart';
 import 'src/settings_screen.dart';
 import 'src/utils/auth_utils.dart';
-import 'src/utils/configs.dart' as config;
 import 'src/utils/consts.dart';
 import 'src/utils/platform_utils.dart' as platform_utils;
 import 'src/utils/pref_util.dart';
