+++ chat_sample/lib/src/config.dart
