+++ conf_call_sample/lib/src/utils/platform_configuration_noweb.dart
@@ -1,4 +1,4 @@
-import 'configs.dart';
+import '../config.dart';
 
 void configureNavigation() {}
