+++ conf_call_sample/lib/src/utils/users_utils.dart
@@ -1,5 +1,6 @@
-import 'package:conf_call_sample/src/utils/configs.dart';
 import 'package:connectycube_sdk/connectycube_sdk.dart';
+
+import '../config.dart';
 
 Future<String> getUserNameCached(int userId) async {
   var cachedUser = users.where((user) => user.id == userId).firstOrNull;
