+++ conf_call_sample/pubspec.yaml
@@ -1,7 +1,7 @@
 name: conf_call_sample
 description: Conf Calls Sample
 
-version: 0.4.1
+version: 0.5.0
 
 environment:
   sdk: '>=3.0.0 <4.0.0'
@@ -11,13 +11,13 @@ dependencies:
     sdk: flutter
 
   assets_audio_player: ^3.1.1
-  connectycube_sdk: ^2.13.1
+  connectycube_sdk: ^3.0.1
 #  connectycube_sdk:
 #    path: ../../flutter-sdk
   connectycube_flutter_call_kit: ^2.7.0
 #  connectycube_flutter_call_kit:
 #    path: ../../connectycube-flutter-call-kit/connectycube_flutter_call_kit
-  device_info_plus: ^10.1.0
+  device_info_plus: ^10.1.2
   flutter_background: ^1.2.0
   flutter_speed_dial: ^7.0.0
   package_info_plus: ^8.0.3
