+++ p2p_call_sample/pubspec.yaml
@@ -1,7 +1,7 @@
 name: p2p_call_sample
 description: P2P Calls Sample
 
-version: 1.7.0
+version: 1.8.0
 
 environment:
   sdk: '>=2.12.0 <3.0.0'
@@ -10,7 +10,7 @@ dependencies:
   flutter:
     sdk: flutter
 
-  connectycube_sdk: ^2.13.0
+  connectycube_sdk: ^3.0.1
 #  connectycube_sdk:
 #    path: ../../flutter-sdk
   shared_preferences: ^2.0.15
@@ -22,7 +22,7 @@ dependencies:
   flutter_background: ^1.3.0
   flutter_speed_dial: ^6.2.0
   permission_handler: ^10.0.1
-  device_info_plus: ^9.0.0
+  device_info_plus: ^10.1.2
   package_info_plus: ^4.0.0
 
 dev_dependencies:
